feat: add bundle analyzer to example
fix: more refactoring to facilitate lazy load and tree-shaking
