+++ packages/plugins/content/image/src/Component/index.tsx
@@ -21,7 +21,10 @@
  */
 
 import * as React from 'react';
-import { ImageLoaded, ImageUploaded } from '@react-page/ui/lib/ImageUpload';
+import {
+  ImageLoaded,
+  ImageUploaded
+} from '@react-page/ui/lib/ImageUpload/types';
 import { ImageProps } from '../types/component';
 import { ImageState } from './../types/state';
