+++ packages/plugins/content/image/src/types/renderer.ts
@@ -1,5 +1,5 @@
 import { ImageProps } from './component';
-import { ImageLoaded } from '@react-page/ui/lib/ImageUpload';
+import { ImageLoaded } from '@react-page/ui/lib/ImageUpload/types';
 
 export interface ImageRendererExtraProps {
   imagePreview?: ImageLoaded;
