+++ packages/plugins/content/slate/src/serialization/createInitialState.ts
@@ -1,0 +1,25 @@
+import { Value, BlockJSON } from 'slate';
+import { P } from '../plugins/paragraph';
+
+export default () => ({
+  editorState: Value.fromJSON({
+    document: {
+      nodes: [
+        {
+          object: 'block',
+          type: P,
+          nodes: [
+            {
+              object: 'text',
+              leaves: [
+                {
+                  text: '',
+                },
+              ],
+            },
+          ],
+        } as BlockJSON,
+      ],
+    },
+  }),
+});
