+++ packages/plugins/layout/background/src/Controls/sub/Color.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { ColorPicker } from '@react-page/ui';
-import { RGBColor } from '@react-page/ui/lib/ColorPicker';
+import { RGBColor } from '@react-page/ui/lib/ColorPicker/types';
 import { BackgroundProps } from '../../types/component';
 
 export interface ColorComponentProps {
