+++ packages/plugins/layout/background/src/Controls/sub/LinearGradient.tsx
@@ -6,7 +6,7 @@ import IconButton from '@material-ui/core/IconButton';
 import DeleteIcon from '@material-ui/icons/Delete';
 
 import { ColorPicker } from '@react-page/ui';
-import { RGBColor } from '@react-page/ui/lib/ColorPicker';
+import { RGBColor } from '@react-page/ui/lib/ColorPicker/types';
 import { BackgroundProps } from '../../types/component';
 
 export interface LinearGradientComponentProps {
@@ -207,7 +207,8 @@ class LinearGradientComponent extends React.Component<
               </div>
               <div>
                 <Typography variant="body1" id="linear-gradient-opacity-label">
-                  {this.props.translations.gradientOpacity} ({(opacity * 100).toFixed(0)}
+                  {this.props.translations.gradientOpacity} (
+                  {(opacity * 100).toFixed(0)}
                   %)
                 </Typography>
                 <Slider
