+++ packages/plugins/layout/background/src/default/settings.tsx
@@ -1,7 +1,11 @@
 import * as React from 'react';
 import { BackgroundSettings } from '../types/settings';
-import { IMAGE_MODE_FLAG, COLOR_MODE_FLAG, GRADIENT_MODE_FLAG } from './../const/mode';
-import { defaultTranslations as defaultImageUploadTranslations } from '@react-page/ui/lib/ImageUpload';
+import {
+  IMAGE_MODE_FLAG,
+  COLOR_MODE_FLAG,
+  GRADIENT_MODE_FLAG
+} from './../const/mode';
+import { defaultTranslations as defaultImageUploadTranslations } from '@react-page/ui/lib/ImageUpload/defaultTranslations';
 
 export const defaultTranslations = {
   ...defaultImageUploadTranslations,
@@ -33,6 +37,6 @@ export const defaultSettings: Partial<BackgroundSettings> = {
   defaultIsParallax: true,
   translations: defaultTranslations,
   enabledModes: IMAGE_MODE_FLAG | COLOR_MODE_FLAG | GRADIENT_MODE_FLAG,
-  Controls: () => <> Controls for this plugin were not provided</> ,
+  Controls: () => <> Controls for this plugin were not provided</>,
   Renderer: () => <>Renderer; for this plugin was not provided </>,
 };
