+++ packages/plugins/layout/background/src/types/api.ts
@@ -1,6 +1,6 @@
 import { ModeEnum } from './modeEnum';
-import { RGBColor } from '@react-page/ui/lib/ColorPicker';
-import { ImageLoaded } from '@react-page/ui/lib/ImageUpload';
+import { RGBColor } from '@react-page/ui/lib/ColorPicker/types';
+import { ImageLoaded } from '@react-page/ui/lib/ImageUpload/types';
 
 export interface BackgroundApi {
   handleChangeDarken: () => void;
