+++ packages/plugins/layout/background/src/types/settings.ts
@@ -1,8 +1,8 @@
 import { BackgroundRendererProps } from './renderer';
 import { BackgroundControlsProps } from './controls';
 import { ContentPluginConfig } from '@react-page/core/lib/service/plugin/classes';
-import { RGBColor } from '@react-page/ui/lib/ColorPicker';
-import { ImageUploadType } from '@react-page/ui/lib/ImageUpload';
+import { RGBColor } from '@react-page/ui/lib/ColorPicker/types';
+import { ImageUploadType } from '@react-page/ui/lib/ImageUpload/types';
 import { ModeEnum } from './modeEnum';
 import { Translations } from './translations';
