+++ packages/ui/src/ColorPicker/colorToString.tsx
@@ -1,0 +1,3 @@
+import { RGBColor } from './types';
+export const colorToString = (c: RGBColor) =>
+  c && 'rgba(' + c.r + ', ' + c.g + ', ' + c.b + ', ' + c.a + ')';
