+++ packages/ui/src/ImageUpload/defaultTranslations.tsx
@@ -1,0 +1,8 @@
+export const defaultTranslations = {
+  buttonContent: 'Upload image',
+  noFileError: 'No file selected',
+  badExtensionError: 'Bad file type',
+  tooBigError: 'Too big',
+  uploadingError: 'Error while uploading',
+  unknownError: 'Unknown error',
+};
